modifico diagrama, hago estructura correcta del MVC, logica en emparejamiento, partido y estados
