--- TP-procesoDesarrolloDeSoftware/src/modelo/Deporte.java
@@ -1,6 +1,0 @@
-package modelo;
-
-public class Deporte {
-    private String nombre;
-    private int cantJugadores;
-}
