--- TP-procesoDesarrolloDeSoftware/src/modelo/EstadoParticipacion.java
@@ -1,7 +1,0 @@
-package modelo;
-
-public enum EstadoParticipacion {
-    PENDIENTE,
-    ACEPTADO,
-    RECHAZADO
-}
