--- TP-procesoDesarrolloDeSoftware/src/modelo/Jugador.java
@@ -1,11 +1,0 @@
-package modelo;
-
-public class Jugador {
-    private String nombreUsuario;
-    private String email;
-    private String contrasenia;
-    private Deporte deporteFavorito;
-    private NivelJuego nivelJuego;
-    private String celular;
-    private Ubicacion ubicacion;
-}
