--- TP-procesoDesarrolloDeSoftware/src/modelo/NivelJuego.java
@@ -1,7 +1,0 @@
-package modelo;
-
-public enum NivelJuego {
-    PRINCIPIANTE,
-    INTERMEDIO,
-    AVANZADO
-}
