--- TP-procesoDesarrolloDeSoftware/src/modelo/ParticipacionJugadorPartido.java
@@ -1,4 +1,0 @@
-package modelo;
-
-public class ParticipacionJugadorPartido {
-}
