--- TP-procesoDesarrolloDeSoftware/src/modelo/Ubicacion.java
@@ -1,4 +1,0 @@
-package modelo;
-
-public class Ubicacion {
-}
