+++ TP-procesoDesarrolloDeSoftware/src/modelo/adapter/AdapterEmailJavaMail.java
@@ -1,0 +1,10 @@
+package modelo.adapter;
+
+import modelo.entidad.notificacion.Notificacion;
+
+public class AdapterEmailJavaMail implements IAdapterEmail{
+    @Override
+    public void enviarEmail(Notificacion notificacion){
+
+    }
+}
