+++ TP-procesoDesarrolloDeSoftware/src/modelo/adapter/IAdapterEmail.java
@@ -1,0 +1,7 @@
+package modelo.adapter;
+
+import modelo.entidad.notificacion.Notificacion;
+
+public interface IAdapterEmail {
+    void enviarEmail(Notificacion notificacion);
+}
