--- TP-procesoDesarrolloDeSoftware/src/modelo/emparejamiento/Emparejador.java
@@ -1,13 +1,0 @@
-package modelo.emparejamiento;
-
-public class Emparejador {
-    private IEmparejador estrategiaEmparejamiento;
-
-    public void emparejar() {
-
-    }
-
-    public void cambiarEstrategiaEmparejamiento(IEmparejador estrategia) {
-        this.estrategiaEmparejamiento = estrategia;
-    }
-}
