--- TP-procesoDesarrolloDeSoftware/src/modelo/emparejamiento/EmparejamientoHistorial.java
@@ -1,8 +1,0 @@
-package modelo.emparejamiento;
-
-public class EmparejamientoHistorial implements IEmparejador{
-    @Override
-    public void emparejar() {
-
-    }
-}
