--- TP-procesoDesarrolloDeSoftware/src/modelo/emparejamiento/EmparejamientoNivel.java
@@ -1,8 +1,0 @@
-package modelo.emparejamiento;
-
-public class EmparejamientoNivel implements IEmparejador{
-    @Override
-    public void emparejar() {
-
-    }
-}
