--- TP-procesoDesarrolloDeSoftware/src/modelo/emparejamiento/EmparejamientoUbicacion.java
@@ -1,8 +1,0 @@
-package modelo.emparejamiento;
-
-public class EmparejamientoUbicacion implements IEmparejador{
-    @Override
-    public void emparejar() {
-
-    }
-}
