--- TP-procesoDesarrolloDeSoftware/src/modelo/emparejamiento/IEmparejador.java
@@ -1,5 +1,0 @@
-package modelo.emparejamiento;
-
-public interface IEmparejador {
-    void emparejar();
-}
