+++ TP-procesoDesarrolloDeSoftware/src/modelo/entidad/Deporte.java
@@ -1,0 +1,6 @@
+package modelo.entidad;
+
+public class Deporte {
+    private String nombre;
+    private int cantJugadores;
+}
