+++ TP-procesoDesarrolloDeSoftware/src/modelo/entidad/EstadoParticipacion.java
@@ -1,0 +1,7 @@
+package modelo.entidad;
+
+public enum EstadoParticipacion {
+    PENDIENTE,
+    ACEPTADO,
+    RECHAZADO
+}
