+++ TP-procesoDesarrolloDeSoftware/src/modelo/entidad/ParticipacionJugadorPartido.java
@@ -1,0 +1,4 @@
+package modelo.entidad;
+
+public class ParticipacionJugadorPartido {
+}
