+++ TP-procesoDesarrolloDeSoftware/src/modelo/entidad/jugador/NivelJuego.java
@@ -1,0 +1,7 @@
+package modelo.entidad.jugador;
+
+public enum NivelJuego {
+    PRINCIPIANTE,
+    INTERMEDIO,
+    AVANZADO
+}
