+++ TP-procesoDesarrolloDeSoftware/src/modelo/entidad/jugador/Ubicacion.java
@@ -1,0 +1,4 @@
+package modelo.entidad.jugador;
+
+public class Ubicacion {
+}
