+++ TP-procesoDesarrolloDeSoftware/src/modelo/entidad/partido/Resenia.java
@@ -1,0 +1,14 @@
+package modelo.entidad.partido;
+
+public class Resenia {
+    private int puntuacion;
+    private String conmentario;
+
+    public void setConmentario(String conmentario) {
+        this.conmentario = conmentario;
+    }
+
+    public void setPuntuacion(int puntuacion) {
+        this.puntuacion = puntuacion;
+    }
+}
