--- TP-procesoDesarrolloDeSoftware/src/modelo/notificacion/AdapterEmailJavaMail.java
@@ -1,8 +1,0 @@
-package modelo.notificacion;
-
-public class AdapterEmailJavaMail implements IAdapterEmail{
-    @Override
-    public void enviarEmail(Notificacion notificacion){
-
-    }
-}
