--- TP-procesoDesarrolloDeSoftware/src/modelo/notificacion/IAdapterEmail.java
@@ -1,5 +1,0 @@
-package modelo.notificacion;
-
-public interface IAdapterEmail {
-    void enviarEmail(Notificacion notificacion);
-}
