--- TP-procesoDesarrolloDeSoftware/src/modelo/notificacion/IServicioNotificacion.java
@@ -1,9 +1,0 @@
-package modelo.notificacion;
-
-import modelo.Jugador;
-
-import java.util.Set;
-
-public interface IServicioNotificacion {
-    void notificar(Notificacion notificacion, Set<Jugador> jugadores);
-}
