--- TP-procesoDesarrolloDeSoftware/src/modelo/notificacion/NotificacionPush.java
@@ -1,13 +1,0 @@
-package modelo.notificacion;
-
-import modelo.Jugador;
-
-import java.util.Set;
-
-public class NotificacionPush implements IServicioNotificacion{
-
-    @Override
-    public void notificar(Notificacion notificacion, Set<Jugador> jugadores) {
-
-    }
-}
