--- TP-procesoDesarrolloDeSoftware/src/modelo/partido/IEstadoPartido.java
@@ -1,10 +1,0 @@
-package modelo.partido;
-
-public interface IEstadoPartido {
-    void cancelar(Partido partido);
-    void necesitarJugadores(Partido partido);
-    void armar(Partido partido);
-    void confirmar(Partido partido);
-    void jugar(Partido partido);
-    void finalizar(Partido partido);
-}
