--- TP-procesoDesarrolloDeSoftware/src/modelo/partido/PartidoCancelado.java
@@ -1,33 +1,0 @@
-package modelo.partido;
-
-public class PartidoCancelado implements IEstadoPartido{
-    @Override
-    public void cancelar(Partido partido) {
-
-    }
-
-    @Override
-    public void necesitarJugadores(Partido partido) {
-
-    }
-
-    @Override
-    public void armar(Partido partido) {
-
-    }
-
-    @Override
-    public void confirmar(Partido partido) {
-
-    }
-
-    @Override
-    public void jugar(Partido partido) {
-
-    }
-
-    @Override
-    public void finalizar(Partido partido) {
-
-    }
-}
