--- TP-procesoDesarrolloDeSoftware/src/modelo/partido/Resenia.java
@@ -1,14 +1,0 @@
-package modelo.partido;
-
-public class Resenia {
-    private int puntuacion;
-    private String conmentario;
-
-    public void setConmentario(String conmentario) {
-        this.conmentario = conmentario;
-    }
-
-    public void setPuntuacion(int puntuacion) {
-        this.puntuacion = puntuacion;
-    }
-}
