+++ TP-procesoDesarrolloDeSoftware/src/modelo/state/PartidoConfirmado.java
@@ -1,0 +1,35 @@
+package modelo.state;
+
+import modelo.entidad.partido.Partido;
+
+public class PartidoConfirmado implements IEstadoPartido {
+    @Override
+    public void cancelar(Partido partido) {
+
+    }
+
+    @Override
+    public void necesitarJugadores(Partido partido) {
+
+    }
+
+    @Override
+    public void armar(Partido partido) {
+
+    }
+
+    @Override
+    public void confirmar(Partido partido) {
+
+    }
+
+    @Override
+    public void jugar(Partido partido) {
+
+    }
+
+    @Override
+    public void finalizar(Partido partido) {
+
+    }
+}
