+++ TP-procesoDesarrolloDeSoftware/src/modelo/strategy/notificacion/NotificacionPush.java
@@ -1,0 +1,14 @@
+package modelo.strategy.notificacion;
+
+import modelo.entidad.jugador.Jugador;
+import modelo.entidad.notificacion.Notificacion;
+
+import java.util.Set;
+
+public class NotificacionPush implements IServicioNotificacion{
+
+    @Override
+    public void notificar(Notificacion notificacion, Set<Jugador> jugadores) {
+
+    }
+}
